+++ frontend/public/index.html
@@ -6,7 +6,7 @@
     <meta name="theme-color" content="#000000" />
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
     <script src="https://kit.fontawesome.com/012219e772.js" crossorigin="anonymous"></script>
-    <title>React Template</title>
+    <title>AirBnb-Clone</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
