+++ frontend/src/components/Navigation/ProfileButton.js
@@ -1,9 +1,32 @@
-const Bed = () => {
-    return (
-      <div style={{ color: "pink", fontSize: "100px" }}>
-        <i class="fa-solid fa-bed"></i>
-      </div>
-    );
+import React, { useState } from "react";
+import { useDispatch } from 'react-redux';
+import * as sessionActions from '../../store/session';
+
+function ProfileButton({ user }) {
+  const dispatch = useDispatch();
+
+  const logout = (e) => {
+    e.preventDefault();
+    dispatch(sessionActions.logout());
   };
 
-  export default Bed;
+  const ulClassName = "profile-dropdown";
+
+  return (
+    <>
+      <button>
+        <i className="fas fa-user-circle" />
+      </button>
+      <ul className="profile-dropdown">
+        <li>{user.username}</li>
+        <li>{user.firstName} {user.lastName}</li>
+        <li>{user.email}</li>
+        <li>
+          <button onClick={logout}>Log Out</button>
+        </li>
+      </ul>
+    </>
+  );
+}
+
+export default ProfileButton;
